+++ src/js_realm.hpp
@@ -1607,6 +1607,12 @@ realm::Realm::Config RealmClass<T>::write_copy_to_helper(ContextType ctx, Object
     ObjectDefaultsMap defaults;
     ConstructorMap constructors;
     bool schema_updated = get_realm_config(ctx, args.count, args.value, config, defaults, constructors);
+
+    // TODO: remove when https://github.com/realm/realm-core/issues/5798 is fixed
+    if (config.sync_config && config.sync_config->flx_sync_requested) {
+        throw std::runtime_error("`writeCopyTo` does not support flexible sync");
+    }
+
     return config;
 }
 
@@ -1709,6 +1715,11 @@ void RealmClass<T>::writeCopyTo(ContextType ctx, ObjectType this_object, Argumen
     }
     else {
         config = write_copy_to_helper_deprecated(ctx, this_object, args);
+    }
+
+    // TODO: remove when https://github.com/realm/realm-core/issues/5798 is fixed
+    if (config.sync_config && config.sync_config->flx_sync_requested) {
+        throw std::runtime_error("`writeCopyTo` does not support flexible sync");
     }
 
     realm->convert(config);